+++ lib/src/extensions/replace_where.dart
@@ -1,0 +1,14 @@
+extension ReplaceWhereExtension<T> on Iterable<T> {
+  /// Replaces every element that matches the [where] condition with [value].
+  Iterable<T> replaceWhere(T value, {required bool Function(T) where}) sync* {
+    final iterator = this.iterator;
+
+    while (iterator.moveNext()) {
+      if (where(iterator.current)) {
+        yield value;
+      } else {
+        yield iterator.current;
+      }
+    }
+  }
+}